\name{clusterVAR-internal}
\title{Internal ClusterVAR functions}

\alias{checkVarianceY}
\alias{plotHeat}
\alias{DetPredSubj}
\alias{calculateICL}
\alias{calculateBIC}
\alias{InitPseudoRand}
\alias{InitRat}
\alias{reorderLags}
\alias{determineLagOrder}
\alias{createX}
\alias{createOutputList}
\alias{constraintsOnB}
\alias{checkSingularitySigma}
\alias{checkPosteriorsNA}
\alias{checkOutliers}
\alias{checkLikelihoodsNA}
\alias{checkConvergence}
\alias{checkComponentsCollapsed}
\alias{callEMFuncs}
\alias{callCalculateCoefficientsForRandoAndRational}
\alias{calculateW}
\alias{calculateU}
\alias{calculateTau}
\alias{calculateSigma}
\alias{calculateRatio}
\alias{calculatePosterior}
\alias{calculateNPara}
\alias{calculateLagList}
\alias{calculateIC}
\alias{calculateFYZ}
\alias{calculateCoefficientsForRandoAndRational}
\alias{calculateBandWZero}
\alias{calculateB}
\alias{calculateA}
\alias{InitFuncs}
\alias{EMInit}
\alias{EMFunc}

\description{Internal ClusterVAR functions.}

\author{Anja Ernst & Jonas Haslbeck}

\details{

These are internal functions.

}

\value{No return value, called for side effects}

\keyword{internal}
