\name{print.ClusterVARCoef}
\alias{print.ClusterVARCoef}
\title{
Overview of Parameters of given Model.
}
\description{
Overview of Parameters of given Model.}
\usage{
\method{print}{ClusterVARCoef}(x, ...)
}

\arguments{
  \item{x}{
An output object of the \code{coef.ClusterVAR} function.
}
  \item{\dots}{
Pass additional arguments.
}
}

\value{
Prints an overview of the fitted model in the console.
}
\author{
Anja Ernst & Jonas Haslbeck
}
\examples{
# Add example later
}


