\name{print.ClusterVARSummary}
\alias{print.ClusterVARSummary}
\title{
Print Summary of Models into Console.
}
\description{
Print Summary of Models into Console.
}
\usage{
\method{print}{ClusterVARSummary}(x, ...)
}

\arguments{
  \item{x}{
An output object of the \code{summary.ClusterVAR} function.
}
  \item{\dots}{
Pass additional arguments.
}
}

\value{
Prints the summary of the fitted models in the console.
}
\author{
Anja Ernst & Jonas Haslbeck
}
\examples{
# Add example later
}


