\name{print.PredictableObs}
\alias{print.PredictableObs}
\title{
    Print Summary of predictable Observations into Console.
}
\description{
    Print Summary of predictable Observations into Console.
}
\usage{
    \method{print}{PredictableObs}(x, ...)
}

\arguments{
    \item{x}{
        An output object of the \code{numberPredictableObservations} function.
    }
    \item{\dots}{
        Pass additional arguments.
    }
}

\value{
    Prints the summary of the number of predictable Observations in the console.
}


\author{
    Anja Ernst & Jonas Haslbeck
}



