\name{clusterVAR-datasets}
\title{Datasets included in the ClusterVAR package}

\alias{ExampleData}
\alias{SyntheticData}


\description{Internal clusterVAR datasets.}

\author{Anja Ernst & Jonas Haslbeck}

\details{

SyntheticData: Follows the structure where a 'Beep' variable and a 'Day' variable together indicate the time-point of the measurement. This data has missing observations that have been removed.

ExampleData: Follows the structure where a single 'Time' variable indicates the time-point of the measurement. This data is generated from a 2 cluster model wtih 1 lag for both clusters (i.e., a c(1, 1) Model).


}

\keyword{internal}
