\name{clusterVAR-datasets}
\title{Datasets included in the ClusterVAR package}

\alias{ExampleData}
\alias{Data}
\alias{OneClusterDataSet}

\description{Internal vlusterVAR functions.}

\author{Anja Ernst & Jonas Haslbeck}

\details{

Data: XXX DESCRIBE

ExampleData: XXX DESCRIBE

OneClusterDataSet: XXX DESCRIBE

}

\keyword{internal}
