\name{print.ClusterVAR}
\alias{print.ClusterVAR}

\title{
Prints description of ClusterVAR objects
}
\description{
Takes the output of the ClusterVAR object and prints a small overview of the fitted model(s).
}
\usage{
\method{print}{ClusterVAR}(x, ...)
}

\arguments{
  \item{x}{
An output object of the \code{LCVAR} function
}
  \item{\dots}{
Pass additional arguments
}
}
\details{
Prints an overview of the fitted model(s) in the console.
}

\author{
Anja Ernst & Jonas Haslbeck
}

